pyproject.toml
[project]
name = "seezar-challenge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dbos>=1.12.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.1",
    "langfuse>=3.3.2",
    "neo4j>=5.28.2",
    "pgvector>=0.4.1",
    "psycopg>=3.2.9",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "reflex>=0.8.8",
    "requests>=2.32.5",
    "stagehand>=0.5.2",
]
